Separate code into two classes
