+++ CollapseLauncher/XAMLs/MainApp/Pages/Dialogs/SimpleDialogs.cs
@@ -1,17 +1,18 @@
-using CollapseLauncher.CustomControls;
-using CollapseLauncher.Statics;
-using CommunityToolkit.WinUI;
-using Hi3Helper;
-using Hi3Helper.Preset;
 using Microsoft.UI.Text;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Documents;
 using Microsoft.UI.Xaml.Media;
+using CommunityToolkit.WinUI;
 using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using System.Collections.Generic;
+using Hi3Helper;
+using Hi3Helper.Preset;
+using CollapseLauncher.Statics;
+using CollapseLauncher.ShortcutsUtils;
+using CollapseLauncher.CustomControls;
 using static Hi3Helper.Data.ConverterTool;
 using static Hi3Helper.Locale;
 using static Hi3Helper.Preset.ConfigV2Store;
